+++ config.yml
@@ -1,3 +1,3 @@
 username: 'root'
 password: ''
-database: 'percona_migrator_test'
+database: 'departure_test'
