+++ departure.gemspec
@@ -2,17 +2,17 @@
 lib = File.expand_path('../lib', __FILE__)
 $LOAD_PATH.unshift(lib) unless $LOAD_PATH.include?(lib)
 
-require 'percona_migrator/version'
+require 'departure/version'
 
 Gem::Specification.new do |spec|
-  spec.name          = 'percona_migrator'
+  spec.name          = 'departure'
   spec.version       = PerconaMigrator::VERSION
   spec.authors       = ['Ilya Zayats', 'Pau Pérez', 'Fran Casas', 'Jorge Morante', 'Adrian Serafin']
   spec.email         = ['ilya.zayats@redbooth.com', 'pau.perez@redbooth.com', 'fran.casas@redbooth.com', 'jorge.morante@redbooth.com', 'adrian@softmad.pl']
 
   spec.summary       = %q{pt-online-schema-change runner for ActiveRecord migrations}
   spec.description   = %q{Execute your ActiveRecord migrations with Percona's pt-online-schema-change}
-  spec.homepage      = 'http://github.com/redbooth/percona_migrator'
+  spec.homepage      = 'http://github.com/redbooth/departure'
   spec.license       = 'MIT'
 
   spec.post_install_message = <<-MESSAGE
