+++ departure_error.log
@@ -1,0 +1,1 @@
+No foreign keys reference `percona_migrator_test`.`comments`; ignoring --alter-foreign-keys-method.
