+++ lib/departure.rb
@@ -1,18 +1,18 @@
 require 'active_record'
 require 'active_support/all'
 
-require 'percona_migrator/version'
-require 'percona_migrator/log_sanitizers/password_sanitizer'
-require 'percona_migrator/runner'
-require 'percona_migrator/cli_generator'
-require 'percona_migrator/logger'
-require 'percona_migrator/null_logger'
-require 'percona_migrator/logger_factory'
-require 'percona_migrator/configuration'
-require 'percona_migrator/errors'
-require 'percona_migrator/command'
+require 'departure/version'
+require 'departure/log_sanitizers/password_sanitizer'
+require 'departure/runner'
+require 'departure/cli_generator'
+require 'departure/logger'
+require 'departure/null_logger'
+require 'departure/logger_factory'
+require 'departure/configuration'
+require 'departure/errors'
+require 'departure/command'
 
-require 'percona_migrator/railtie' if defined?(Rails)
+require 'departure/railtie' if defined?(Rails)
 
 # We need the OS not to buffer the IO to see pt-osc's output while migrating
 $stdout.sync = true
