+++ lib/departure/cli_generator.rb
@@ -1,8 +1,8 @@
-require 'percona_migrator/dsn'
-require 'percona_migrator/option'
-require 'percona_migrator/alter_argument'
-require 'percona_migrator/connection_details'
-require 'percona_migrator/user_options'
+require 'departure/dsn'
+require 'departure/option'
+require 'departure/alter_argument'
+require 'departure/connection_details'
+require 'departure/user_options'
 
 module PerconaMigrator
