+++ lib/departure/command.rb
