+++ lib/departure/configuration.rb
@@ -4,7 +4,7 @@ module PerconaMigrator
 
     def initialize
       @tmp_path = '.'.freeze
-      @error_log_filename = 'percona_migrator_error.log'.freeze
+      @error_log_filename = 'departure_error.log'.freeze
       @global_percona_args = nil
     end
