+++ lib/departure/connection_details.rb
