+++ lib/departure/log_sanitizers/password_sanitizer.rb
