+++ lib/departure/logger.rb
