+++ lib/departure/option.rb
