+++ spec/departure/configuration_spec.rb
@@ -3,7 +3,7 @@ require 'spec_helper'
 describe PerconaMigrator::Configuration do
   describe '#initialize' do
     its(:tmp_path) { is_expected.to eq('.') }
-    its(:error_log_filename) { is_expected.to eq('percona_migrator_error.log') }
+    its(:error_log_filename) { is_expected.to eq('departure_error.log') }
   end
 
   describe '#tmp_path' do
