+++ spec/departure/connection_details_spec.rb
