+++ spec/departure/log_sanitizers/password_sanitizer_spec.rb
