+++ spec/departure/logger_factory_spec.rb
