+++ spec/departure/runner_spec.rb
@@ -11,7 +11,7 @@ describe PerconaMigrator::Runner do
   let(:config) do
     instance_double(
       PerconaMigrator::Configuration,
-      error_log_path: 'percona_migrator_error.log'
+      error_log_path: 'departure_error.log'
     )
   end
